items get command
