+++ zbx.py
@@ -65,8 +65,8 @@ config = configparser.ConfigParser()
 config.read(configfile)
 try:
     ZABBIX_HTTP_AUTH = config['zabbix']['http_auth']
-except :
-    warn = eLIGHT_RED +'WARN : ' + eCOLOR_NONE + 'check your configini file because key http_auth not exist'
+except:
+    warn = eLIGHT_RED + 'WARN : ' + eCOLOR_NONE + 'check your configini file because key http_auth not exist'
     click.echo(warn)
     click.pause()
     ZABBIX_HTTP_AUTH = 'false'
@@ -131,6 +131,11 @@ def alert():
     pass
 
 
+@zabbix.group()
+def items():
+    """Item sub-commands management."""
+    pass
+
 ##########################################################################
 # FUNCTIONS
 ##########################################################################
@@ -172,6 +177,23 @@ def get_group_id(group_name):
         return "not found"
     else:
         result = response[0]["groupid"]
+        return result
+
+
+def get_host_in_a_group(group_id):
+    """Get list of host in a group."""
+    response = zapi.host.get(
+        output='extend',
+        groupids=group_id,
+        monitored_hosts=1,
+        sortfield='host'
+    )
+    response.sort()
+    if not response:
+        click.echo('No host found in group id "%s"' % group_id)
+        sys.exit(1)
+    else:
+        result = response
         return result
 
 
@@ -252,6 +274,44 @@ def get_user_id(user_alias):
         return "not found"
     else:
         return response[0]["userid"]
+
+
+def get_key_item(host_id, item_key_name):
+    """Get the key item id."""
+    response = zapi.item.get(
+        output="extend",
+        hostids=host_id,
+        search={"key_": item_key_name},
+        sortfield="name"
+    )
+    if not response:
+        click.echo('item key name "%s" not found in host id "%s"' % (item_key_name, host_id))
+        sys.exit(1)
+    else:
+        result = response
+        # import pdb; pdb.set_trace()
+        return result
+
+
+def get_item_last_value(item_value_type, item_searched_id, host_id):
+    """Get item last value."""
+    response = zapi.history.get(
+        history=item_value_type,
+        output='extend',
+        itemids=item_searched_id,
+        time_from=time.time() - 7200,
+        time_till=time.time(),
+        hostids=host_id,
+        sortfield='clock',
+        sortorder='DESC',
+        limit=1
+    )
+    if not response:
+        click.echo('No item value found : %s %s %s' % (item_value_type, item_searched_id, host_id))
+        sys.exit(1)
+    else:
+        result = response
+        return result[0]["value"]
 
 ##########################################################################
 # Maintenance sub command
@@ -662,6 +722,52 @@ def unmonitored():
         tableau_unmonitored.append([host["name"], "Not monitored"])
     header = [Bold + "NAME", "STATUS" + UnBold]
     print(tabulate(tableau_unmonitored, headers=header, tablefmt="plain"))
+
+
+##########################################################################
+# zbx item COMMANDS
+##########################################################################
+
+@items.command("get")
+@click.argument('item_key_name')
+@click.argument('source_name')
+@click.option('--group', 'source', flag_value='group', default=True)
+@click.option('--host', 'source', flag_value='host')
+def items_get(item_key_name, source_name, source):
+    """List last value for a item key for all host in a groupe name."""
+    if source == "group":
+        group_id = get_group_id(source_name)
+        tableau_server_in_group = []
+
+        # List host in group
+        list_servers = get_host_in_a_group(group_id)
+
+        # get the id of the firt host in the list and
+        # get the value_type of the item key required
+        # # import pdb; pdb.set_trace()
+        firt_host_id = get_host_id(list_servers[0]['host'])
+        item = get_key_item(firt_host_id, item_key_name)
+        item_value_type = item[0]['value_type']
+
+        # For eatch host
+        # Find the host id
+        # Find the item id
+        # Get the lastdata for the item
+        for host in list_servers:
+            host_id = get_host_id(host)
+            item = get_key_item(host_id, item_key_name)
+            item_searched_id = item[0]['itemid']
+            item_value = get_item_last_value(item_value_type, item_searched_id, host_id)
+            tableau_server_in_group.append([host["name"], source_name, item_value])
+        header = [Bold + "NAME", "Group", item_key_name + UnBold]
+        print(tabulate(tableau_server_in_group, headers=header, tablefmt="plain", numalign="left"))
+    if source == "host":
+        host_id = get_host_id(source_name)
+        item = get_key_item(host_id, item_key_name)
+        item_searched_id = item[0]['itemid']
+        item_value_type = item[0]['value_type']
+        item_value = get_item_last_value(item_value_type, item_searched_id, host_id)
+        click.echo('%s %s %s' % (source_name, item_key_name, item_value))
 
 
 # MAIN
